+++ mitmproxy-rs/Cargo.toml
@@ -18,7 +18,7 @@ anyhow = { version = "1.0.86", features = ["backtrace"] }
 data-encoding = "2.6.0"
 log = "0.4.22"
 once_cell = "1"
-pyo3 = { version = "0.21", features = ["abi3", "abi3-py310", "extension-module", "anyhow"] }
+pyo3 = { version = "0.21", features = ["abi3", "abi3-py310", "extension-module", "anyhow", "experimental-declarative-modules"] }
 pyo3-asyncio-0-21 = { version = "0.21", features = ["tokio-runtime"] }
 pyo3-log = "0.11.0"
 rand_core = { version = "0.6.4", features = ["getrandom"] }
