+++ mitmproxy-rs/mitmproxy_rs/certs.pyi
@@ -1,0 +1,2 @@
+def add_cert(pem: str) -> None: ...
+def remove_cert() -> None: ...
