+++ mitmproxy-rs/mitmproxy_rs/process_info.pyi
@@ -1,0 +1,15 @@
+from pathlib import Path
+from typing import final
+
+def active_executables() -> list[Process]: ...
+def executable_icon(path: Path | str) -> bytes: ...
+@final
+class Process:
+    @property
+    def executable(self) -> str: ...
+    @property
+    def display_name(self) -> str: ...
+    @property
+    def is_visible(self) -> bool: ...
+    @property
+    def is_system(self) -> bool: ...
